pyproject.toml
[project]
name = "ml-wine-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-storage-blob>=12.26.0",
    "certifi~=2023.7.22",
    "chardet~=5.2.0",
    "dotenv>=0.9.9",
    "feature-engine>=1.0.2,<1.6.0",
    "flask~=3.0.3",
    "jinja2~=3.1.4",
    "joblib==1.4.2",
    "markupsafe~=2.1.5",
    "matplotlib~=3.9.0",
    "numpy==1.26.4",
    "packaging~=24.1",
    "pandas>=2.1.2",
    "pillow~=10.4.0",
    "pip>=23.3.0",
    "platformdirs~=4.2.2",
    "pydantic~=2.8.2",
    "pydantic-settings>=2.10.1",
    "pyparsing~=3.1.2",
    "pytest~=8.3.2",
    "pytz~=2023.3.post1",
    "scikit-learn==1.5.1",
    "scipy==1.14.0",
    "seaborn~=0.13.2",
    "setuptools~=68.1.2",
    "six~=1.16.0",
    "strictyaml~=1.7.3",
    "threadpoolctl==3.5.0",
    "tree~=0.2.4",
    "werkzeug~=3.0.3",
]
